# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы для мастеров ручной работы",
        "title": "CraftPlace",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Возвращает список категорий с возможностью фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить категории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию категории",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.CategoryResponse"
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id_category}": {
            "get": {
                "description": "Возвращает информацию о категории по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить категорию по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID категории",
                        "name": "id_category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о категории",
                        "schema": {
                            "$ref": "#/definitions/reqresp.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID категории",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "Возвращает список постов с возможностью фильтрации по магазину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить посты",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Фильтр по ID магазина",
                        "name": "id_shop",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список постов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.PostResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID магазина",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Возвращает список товаров с возможностью фильтрации по различным параметрам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить товары",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию товара",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Минимальная цена товара",
                        "name": "min_cost",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100000,
                        "description": "Максимальная цена товара",
                        "name": "max_cost",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID магазина",
                        "name": "id_shop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID категории",
                        "name": "id_category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.ProductResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/shops": {
            "get": {
                "description": "Возвращает список магазинов с возможностью фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить магазины",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию магазина",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID пользователя",
                        "name": "id_user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.ShopResponse"
                            }
                        }
                    }
                }
            }
        },
        "/shops/{id_shop}": {
            "get": {
                "description": "Возвращает информацию о магазине по его идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить магазин по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID магазина",
                        "name": "id_shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о магазине",
                        "schema": {
                            "$ref": "#/definitions/reqresp.ShopResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID магазина",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "reqresp.CategoryResponse": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "example": "Eco"
                }
            }
        },
...
        "reqresp.PostResponse": {
            "type": "object",
            "required": [
                "description",
                "shopID"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "timePublication": {
                    "type": "string",
                    "example": "2023-06-15T10:00:00Z"
                }
            }
        },
        "reqresp.ProductResponse": {
            "type": "object",
            "required": [
                "categoryIDs",
                "cost",
                "description",
                "shopID"
            ],
            "properties": {
                "categoryIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cost": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 200
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "example": "Eco"
                }
            }
        },
...
        "reqresp.ShopResponse": {
            "type": "object",
            "required": [
                "description",
                "userID"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id_shop": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "example": "Eco"
                },
                "userID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
...
    }
}